apparently, all eslint plugins must be prefixed with "eslint-plugin-..."
https://eslint.org/docs/latest/use/configure/plugins

This was not clear, I'm still not entirely sure that its necessary.
But when trying to use this rule in another project, I can't seem to opt out of the prefix.
Again, eslint continues to make my life hard.

Also, bumped version to 2.0.0 (as this is technically a breaking change)